Feature/restore password

+++ BackEnd/KuroOpti.Common/DTO/RouteDto.cs
@@ -5,10 +5,10 @@ namespace KuroOpti.Common.DTO
         public int Id { get; set; }
         public string? Name { get; set; }
 
-        public decimal StartLat { get; set; }
-        public decimal StartLng { get; set; }
-        public decimal EndLat { get; set; }
-        public decimal EndLng { get; set; }
+        public double StartLat { get; set; }
+        public double StartLng { get; set; }
+        public double EndLat { get; set; }
+        public double EndLng { get; set; }
 
         public string Polyline { get; set; } = default!;
     }
